+++ CHANGELOG.md
@@ -5,3 +5,13 @@
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Created FileSize action
+- Created BulkFileListener
+
+### Changed
+- Updated build.gradle.kts
+- Updated gradle.properties
+- Updated READMe.md
+
+### Removed
+- Removed templatecode